+++ src/l1j/jrwz/server/datatables/LightSpawnTable.java
@@ -79,7 +79,7 @@ public class LightSpawnTable {
                     field.setId(IdFactory.getInstance().nextId());
                     field.setX(rs.getInt("locx"));
                     field.setY(rs.getInt("locy"));
-                    field.setMap((short) rs.getInt("mapid"));
+                    field.setMap( rs.getInt("mapid"));
                     field.setHomeX(field.getX());
                     field.setHomeY(field.getY());
                     field.setHeading(0);
