+++ src/l1j/jrwz/server/model/ElementalStoneGenerator.java
@@ -105,7 +105,7 @@ public class ElementalStoneGenerator implements Runnable {
     public void run() {
         try {
             L1Map map = L1WorldMap.getInstance().getMap(
-                    (short) ELVEN_FOREST_MAPID);
+                    ELVEN_FOREST_MAPID);
             while (true) {
                 removeItemsPickedUp();
